requirements.txt
absl-py==2.3.1
attrs==25.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dm-tree==0.1.9
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
h5py==3.14.0
Jinja2==3.1.6
joblib==1.5.1
keras-core==0.1.7
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
namex==0.1.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
rich==14.0.0
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
wrapt==1.17.2
